mv median crossings to enable downgrade in cxx std req
