+++ include/AbcSmc/AbcUtil.h
@@ -62,10 +62,6 @@ namespace ABC {
 
     float_type skewness(const Col & data);
 
-    float_type median_crossings(const Col & data, const float_type _median);
-
-    float_type median_crossings(const Col & data);
-
     float_type optimize_box_cox(const Col & data, const float lambda_min, const float lambda_max, const float step);
 
     float_type optimize_box_cox(const Col & data);
