+++ scratch/SimUtils.h
@@ -1,0 +1,5 @@
+
+#include <PLS/pls.h>
+
+float_type median_crossings(const Col & data, const float_type _median);
+float_type median_crossings(const Col & data);
